+++ pipetune/plugin/__init__.py
@@ -1,0 +1,2 @@
+"""PipeTune local plugin tooling."""
+